Oops, forgot to stage updated file.  Previous commit message applies here. Copied below:
Created a file mapping the output file paths from the gag file to the corresponding subbasin number.  Note that if you look at the gag file there are currently outputs for 13 "gauges" but I've only included 12 of them in the mapping file.  This is because only the first 12 correspond to actual gauges that we can use for calibration - the 13th was added to check spatial variability of discrepancies between simulated and observed streamflow downstream of Calaveras Reservoir, but there is no gauge located there.